requirements/base.txt
psycopg2
django-mptt
python-decouple
jsonschema
django-redis
hiredis
django-rq
djangorestframework
django
